Add -i (in-place) flag to HSBC script

+++ hsbc.sh
@@ -1,10 +1,20 @@
 #!/bin/bash
 
 script="$(dirname "$0")/hsbc.awk"
+inplace=0
+
+if [ "$1" = "-i" ]; then
+    inplace=1
+    shift
+fi
 
 for file in "$@"; do
     if [ -e "$file" ]; then
-        { rm "$file" && awk -f "$script" > "$file"; } < "$file"
+        if [ $inplace -eq 1 ]; then
+            { rm "$file" && awk -f "$script" > "$file"; } < "$file"
+        else
+            awk -f "$script" "$file"
+        fi
     else
         echo "$file: No such file or directory" >&2
         exit 1
